refactor: introduce new global functions structure
