+++ integration-libs/opf/global-functions/core/services/domains/global/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2026 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './opf-global-functions-global-domain-registrations.service';
+export * from './opf-global-functions-global-domain.service';
