+++ integration-libs/opf/global-functions/core/services/domains/redirect/index.ts
@@ -1,0 +1,8 @@
+/*
+ * SPDX-FileCopyrightText: 2026 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './opf-global-functions-redirect-domain-registrations.service';
+export * from './opf-global-functions-redirect-domain.service';
