+++ integration-libs/opf/global-functions/core/services/index.ts
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: 2026 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './domains';
+export * from './opf-global-functions-shared.service';
+export * from './opf-global-functions-shared-registrations.service';
